feat: `rocks debug project`
